refactor(feishu): extract audio transcription failure resolver
